home-work/pyproject.toml
[project]
name = "home-work"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "langfuse>=3.2.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
